+++ backend/src/interfaces/Dataset.ts
@@ -23,9 +23,9 @@ export interface Dataset {
   data_producer?: string;
   variables_measured?: VariableMeasured[];
   spatial_resolution?: string;
-  publication_date?: Date;
-  reference_period_start?: string;
-  reference_period_stop?: string;
+  publication_date?: string;
+  reference_period_start?: Date;
+  reference_period_stop?: Date;
   licenses?: string[];
   citation?: string;
   geographical_extent?: string;